Add missing const references and use mode descriptive  types

+++ core/include/hash_utils.h
@@ -3,7 +3,7 @@
 #include <Eigen/Eigen>
 
 // this function provides unique index
-unsigned long long int get_index(const int16_t x, const int16_t y, const int16_t z);
+uint64_t get_index(const int16_t x, const int16_t y, const int16_t z);
 
 // this function provides unique index for input point p and 3D space decomposition into buckets b
-unsigned long long int get_rgd_index(const Eigen::Vector3d p, const Eigen::Vector3d b);
+uint64_t get_rgd_index(const Eigen::Vector3d p, const Eigen::Vector3d b);
