+++ core/src/icp.cpp
@@ -8,16 +8,16 @@
 #include <m_estimators.h>
 #include <transformations.h>
 
-std::vector<ICP::Job> ICP::get_jobs(long long unsigned int size, int num_threads)
+std::vector<ICP::Job> ICP::get_jobs(uint64_t size, int num_threads)
 {
     int hc = size / num_threads;
     if (hc < 1)
         hc = 1;
 
     std::vector<Job> jobs;
-    for (long long unsigned int i = 0; i < size; i += hc)
+    for (uint64_t i = 0; i < size; i += hc)
     {
-        long long unsigned int sequence_length = hc;
+        uint64_t sequence_length = hc;
         if (i + hc >= size)
         {
             sequence_length = size - i;
